requirements.txt
python-telegram-bot==21.6
google-genai
python-dotenv==1.0.1
